fix update default group bugs

+++ api/internal/services/oc_user/types.go
@@ -16,7 +16,7 @@ type OcservUserCreateOrUpdateRequest struct {
 	Username    string     `json:"username" validate:"required;min=3,max=16"`
 	Password    string     `json:"password" validate:"required,min=8,max=16"`
 	TrafficType int32      `json:"traffic_type" validate:"required;min=0,max=2"`
-	TrafficSize *int32     `json:"traffic_size"`
+	TrafficSize *float64   `json:"traffic_size"`
 	ExpireAt    *time.Time `json:"expire_at" validate:"required;omitempty"`
 }
